requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
Django==2.0.8
django-cors-headers==2.4.0
django-fcm==0.1.1
django-filter==2.0.0
django-pusherable==0.2.0
django-webpush==0.2.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gcloud==0.17.0
googleapis-common-protos==1.5.5
http-ece==1.0.5
httplib2==0.12.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
jws==0.1.3
kombu==4.2.2.post1
Markdown==2.6.11
MarkupSafe==1.0
ndg-httpsclient==0.5.1
oauth2client==3.0.0
protobuf==3.6.1
pusher==2.1.2
py-vapid==1.4.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.18
pycryptodome==3.4.3
PyJWT==1.6.4
PyNaCl==1.3.0
pyOpenSSL==18.0.0
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2018.5
pywebpush==1.6.0
redis==3.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
vine==1.2.0
